# contracts/public.openapi.yaml
openapi: 3.0.3
info:
  title: VOLA Public API
  version: "1.0.0"
  description: >
    Public API for the VOLA mobile and web clients. See
    docs/architecture/api-conventions.md for the conventions this spec
    follows (versioning, error shape, auth, status codes).


    Two behaviours are applied by response middleware and so are not
    per-operation choices, even though OpenAPI can only express them
    per-operation. Every GET returns a strong `ETag` and honours
    `If-None-Match` with a `304` — both are documented on each GET below.
    Every such response also carries `Cache-Control: private, no-cache` —
    "revalidate before reuse", which is the contract the `ETag` describes.
    A route that opts out sends `Cache-Control: no-store` and no validator
    (`/healthz` does, so a liveness probe is never answered `304`).


    And any response over ~1 KB is gzipped when the client sends
    `Accept-Encoding: gzip`; the `ETag` is computed from the *uncompressed*
    body, so it does not change with the encoding a client asks for.
servers:
  - url: http://localhost:8080/v1
    description: Local development
security:
  - ClerkBearerAuth: []
components:
  securitySchemes:
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued session token, verified server-side via Clerk's JWKS.
  schemas:
...
    Me:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          example: user_3H8rAWz7wri5uEKGajnj4pLqWXz
...
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: >
        An `ETag` from a previous response. If it still matches, the server
        answers `304 Not Modified` with no body. Accepts a comma-separated
        list and `*`.
      schema: { type: string }
...
paths:
...
  /me:
    get:
      summary: Return the authenticated caller's verified identity
      operationId: getMe
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The verified caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /profile:
    get:
      summary: Get the authenticated caller's profile
      operationId: getProfile
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The caller's profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      summary: Create the authenticated caller's profile (onboarding)
      operationId: createProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCreate"
      responses:
        "201":
          description: The newly created profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
    patch:
      summary: Partially update the authenticated caller's profile
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /profile/exercise-units:
    get:
      summary: The caller's per-exercise unit overrides
      operationId: listExerciseUnits
      description: >
        A map of exercise id to unit system. A **missing key means "use the
        profile default"** — there is deliberately no third state, so
        clearing an override removes the key rather than storing a sentinel.

        Exists because a lifter who thinks in kilograms still faces a leg
        press marked in pounds, and converting in your head at the moment
        you're recording a number is exactly what this avoids.
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Overrides only; exercises using the default are absent.
          content:
            application/json:
              schema:
                type: object
                required: [exercise_units]
                properties:
                  exercise_units:
                    type: object
                    additionalProperties:
                      type: string
                      enum: [metric, imperial]
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /profile/exercise-units/{exerciseID}:
    put:
      summary: Set or clear one exercise's unit override
      operationId: setExerciseUnit
      parameters:
        - name: exerciseID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                unit_system:
                  type: string
                  nullable: true
                  enum: [metric, imperial]
                  description: Null clears the override, falling back to the profile default.
      responses:
        "204": { description: "Stored, or cleared." }
        "400":
          description: An unknown exercise, or a unit outside metric/imperial.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /modules:
    get:
      summary: The discipline registry, merged with the caller's own toggles
      description: >
        Every discipline VOLA supports, in display order, each with its
        capabilities and whether this user has it enabled. The catalog and the
        user's choices arrive together because a client needs both before it can
        render navigation — splitting them would mean two requests before the
        nav bar can draw.


        A discipline the caller has never toggled reports its registry default,
        so adding a discipline requires no backfill.
      operationId: listModules
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: All modules with the caller's enablement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleList"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      summary: Toggle one or more modules for the caller
      description: >
        A sparse map — `{"bjj": false}` toggles one module and leaves the rest
        untouched, so two clients editing different toggles cannot clobber each
        other. Returns the full merged set, since the caller's next render needs
        all of it anyway.
      operationId: setModules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: boolean
              minProperties: 1
              example:
                bjj: false
      responses:
        "200":
          description: The full module set after the change.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /flags:
    get:
      summary: List all server-controlled feature flags
      operationId: listFeatureFlags
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: All feature flags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlagList"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /activities:
    get:
      summary: List the authenticated caller's own activities
      operationId: listActivities
      description: >
        The caller's own activity log, newest first. Capped at 500; there is no
        pagination, so older entries are simply not returned.
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The caller's activities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityList"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: >
        Create an activity for the authenticated caller. Idempotent on
        `id` — retrying a sync with the same id returns the original row,
        not an error.
      operationId: createActivity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityCreate"
      responses:
        "200":
          description: The created (or, on a retried id, the original) activity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /exercises:
    get:
      summary: List the global exercise catalog
      operationId: listExercises
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - name: sport
          in: query
          required: false
          schema: { type: string }
          description: Exact-match filter, e.g. `strength` or `running`.
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: Case-insensitive substring match on the name.
      responses:
        "200":
          description: Matching catalog entries, ordered by sport then name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseList"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /exercises/{exerciseID}:
    get:
      summary: Get one catalog entry by its slug
      operationId: getExercise
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - name: exerciseID
          in: path
          required: true
          schema: { type: string }
          example: barbell-back-squat
      responses:
        "200":
          description: The exercise.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /techniques:
    get:
      summary: List the BJJ technique library
      operationId: listTechniques
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - name: position
          in: query
          schema: { type: string }
          example: Guard - Bottom
        - name: category
          in: query
          schema: { type: string }
          example: Submission
        - name: gi
          in: query
          schema: { type: string, enum: [Both, "Gi Only", "No-Gi Only"] }
          description: >
            Asking for gi or no-gi also returns techniques marked `Both` —
            a filter that excluded them would hide most of the library.
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
          description: Case-insensitive substring of the name **or any alias** — "scarf hold" finds "Kesa-Gatame Escape".
      responses:
        "200":
          description: Matching techniques, ordered by position then category then name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TechniqueList" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          description: Invalid filter value.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /techniques/rulesets:
    get:
      summary: List IBJJF rulesets
      operationId: listIBJJFRulesets
      description: >
        All 25 IBJJF rulesets. Fetch once and cache: this is what turns a
        TechniqueSummary's `ibjjf_ruleset_id` into a legality badge without a
        request per row. They change with the IBJJF rulebook, not with the
        library.
      tags: [Techniques]
      security: [{ ClerkBearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Every ruleset, ordered by id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IBJJFRulesetList" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /techniques/positions:
    get:
      summary: List the BJJ position glossary
      operationId: listPositions
      description: >
        All ten positions, already in pedagogical reading order. Fetch once and
        cache, like the rulesets. There is no per-position technique endpoint on
        purpose: a client that already holds the library resolves "techniques
        from here" locally, by prefix-matching each summary's `position`
        against this entry's `family` AND then applying
        `detail_includes`/`detail_excludes` against its `position_detail`.
        Both steps are required — see those fields.
      tags: [Techniques]
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Every position, ordered by order_index.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionList" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /techniques/positions/{positionID}:
    get:
      summary: Get one position
      operationId: getPosition
      tags: [Techniques]
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - name: positionID
          in: path
          required: true
          schema: { type: string }
          example: closed-guard
      responses:
        "200":
          description: The position.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Position" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /techniques/{techniqueID}:
    get:
      summary: Get one technique
      operationId: getTechnique
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - name: techniqueID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The technique.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Technique" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /bjj/standing:
    get:
      summary: The athlete's BJJ rank and promotion history
      description: >
        `current` is DERIVED from `promotions`, not stored — the highest rank
        recorded, which is well defined even when promotions were entered out
        of order or without dates. It is null when nothing has been recorded:
        a new account has no belt, and defaulting to white would put a rank on
        someone who has never trained.
      operationId: getBjjStanding
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Rank and history.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BjjStanding" }
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
        "304":
          $ref: "#/components/responses/NotModified"
        "401": { $ref: "#/components/responses/Unauthorized" }
...